chore: extract cli_variant_struct derive module for reusability in tests
